OC-102: extracted RenderTreeMapJsonActionTest
